Ive added a hashmap for the equipped items, along with some mutators for the player class

+++ Item.java
@@ -2,10 +2,10 @@ public class Item
 {
     private String name;
     private String itemID;
-    private Float weight;
-    private Float size;
+    private double weight;
+    private double size;
 
-    public Item(String name, String itemID, Float weight, Float size)
+    public Item(String name, String itemID, double weight, double size)
     {
         this.name = name;
         this.itemID = itemID;
@@ -17,8 +17,8 @@ public class Item
 
     public String getItemID() { return itemID; }
 
-    public Float getWeight() { return weight; }
+    public double getWeight() { return weight; }
 
-    public Float getSize() { return size; }
+    public double getSize() { return size; }
 
 }
